fix(project-server): rename build/ -> deploy/ (adio ignores build dirs) + sync references
adio treats a directory named `build` as build output and skips it, so it never saw the @vercel/nft
import in src/build/ and flagged the dep as unused. Rename the folder to src/deploy/. Also run
sync-dependencies so cli-core references.json reflects the @vercel/nft move (build-tools -> project-server).
