+++ packages/project-server/src/deploy/CopyExternalDependencies.ts
@@ -1,11 +1,7 @@
 import path from "path";
 import fs from "fs";
 import { nodeFileTrace } from "@vercel/nft";
-import {
-    ApiAfterBuild,
-    GetProjectService,
-    UiService
-} from "@webiny/project/abstractions/index.js";
+import { ApiAfterBuild, GetProjectService, UiService } from "@webiny/project/abstractions/index.js";
 import { getServerBuildPaths } from "./getServerBuildPaths.js";
 
 /**
