+++ packages/project-server/src/deploy/getServerBuildPaths.ts
